Implementa atualização e edição do detalhes verifica se um plano pode ou não ser deletado

+++ app/Http/Controllers/Admin/PlanController.php
@@ -52,14 +52,21 @@ class PlanController extends Controller
     
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository->with('details')
+                ->where('url', $url)->first();
         
         if (!$plan){
             return redirect()->back();
         }
         
+        if ($plan->details->count() > 0){
+            return redirect()->back()
+                    ->with('error', 'Existem detalhes vinculados a esse plano, não pode ser deletado.');
+        }
+        
         $plan->delete();
-        return redirect()->route('plans.index');
+        return redirect()->route('plans.index')
+                ->with('message', 'Registro deletado com sucesso.');
     }
     
     public function search(Request $request)
